uart: Fixed a UB issue that came in with 50984d6a

+++ firmware/src/boards/samd11_xplained/uart.c
@@ -111,10 +111,10 @@ void uart_init(bool configure_pinmux, unsigned long baudrate)
 	// and reduce binary size as a result.
 	// This approach does multiply-and-shift in nested fractions instead.
 	// Exact for every CONF_CPU_FREQUENCY up to 48MHz (max for SAMD11/SAMD21).
-	const uint32_t m1 =  (UINT32_C(1) << 32U) / CONF_CPU_FREQUENCY;
-	const uint32_t m2 = ((UINT32_C(1) << 42U) / CONF_CPU_FREQUENCY) & 0x3FFU;
-	const uint32_t m3 = ((UINT32_C(1) << 52U) / CONF_CPU_FREQUENCY) & 0x3FFU;
-	const uint32_t m4 = ((UINT32_C(1) << 62U) / CONF_CPU_FREQUENCY) & 0x3FFU;
+	const uint32_t m1 =  (UINT64_C(1) << 32U) / CONF_CPU_FREQUENCY;
+	const uint32_t m2 = ((UINT64_C(1) << 42U) / CONF_CPU_FREQUENCY) & 0x3FFU;
+	const uint32_t m3 = ((UINT64_C(1) << 52U) / CONF_CPU_FREQUENCY) & 0x3FFU;
+	const uint32_t m4 = ((UINT64_C(1) << 62U) / CONF_CPU_FREQUENCY) & 0x3FFU;
 	const uint32_t op4 = (baudrate * m4 -  1U) >> 10U;
 	const uint32_t op3 = (baudrate * m3 + op4) >> 10U;
 	const uint32_t op2 = (baudrate * m2 + op3) >> 10U;
